fix windows dtype for faster ernie

+++ examples/experimental/faster_ernie/token_cls/train.py
@@ -87,7 +87,7 @@ def batchify_fn(batch, no_entity_id, ignore_label=-100, max_seq_len=512):
             label += [ignore_label] * (max_seq_len - len(label))
         labels.append(label)
 
-    labels = np.array(labels)
+    labels = np.array(labels, dtype="int64")
     seq_lens = np.array(seq_lens)
     return texts, labels, seq_lens
